drivers: ata: Overflow coverity fix
Typecasting a variable to prevent an overflow.

Coverity ID 24245

Reviewed-on: http://git-master/r/1213345
(cherry picked from commit 15f90540a4c715670d8f76e37bbb5a1dda915ef6)
Reviewed-on: http://git-master/r/1278755
Tested-by: Mikko Perttunen <mperttunen@nvidia.com>

+++ drivers/ata/libata-core.c
@@ -725,7 +725,8 @@ u64 ata_tf_read_block(const struct ata_taskfile *tf, struct ata_device *dev)
 			return U64_MAX;
 		}
 
-		block = (cyl * dev->heads + head) * dev->sectors + sect - 1;
+		block = ((u64)cyl * dev->heads + head) *
+			dev->sectors + sect - 1;
 	}
 
 	return block;
